refactor: move is_typing
